server/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
dashscope>=1.14.0
jieba>=0.42.1
